+++ user.py
@@ -1,8 +1,9 @@
+from flask import request
 from flask_restful import Resource, fields, marshal_with, reqparse, abort
 
 # check if posted data is true
 from models import User, db
-import re
+from cookies import authentication_required, create_token
 
 
 # Create a request parser for dynamically generating arguments based on the desired structure
@@ -10,13 +11,13 @@ def create_user_args():
     user_args = reqparse.RequestParser()
 
     # Add arguments for User model
-    user_args.add_argument('is_premium', type=int, help='Is Premium, required', required=True)
+    user_args.add_argument('is_premium', type=int, help='Is Premium, required', required=False)
     user_args.add_argument('name', type=str, help='Name, required', required=True)
     user_args.add_argument('user_name', type=str, help='User Name, required', required=True)
     user_args.add_argument('chat_id', type=int, help='Chat ID, required', required=True)
-    user_args.add_argument('education_center', type=int, help='Education Center, required', required=True)
+    user_args.add_argument('education_center', type=str, help='Education Center, required', required=True)
     user_args.add_argument('verification', type=int, help='Verification, required', required=True)
-    user_args.add_argument('payment', type=str, help='Payment Information, required', required=True)
+    user_args.add_argument('payment', type=str, help='Payment Information, required', required=False)
 
     return user_args
 
@@ -29,7 +30,7 @@ def update_user_args():
     user_args.add_argument('name', type=str, help='Name', required=False)
     user_args.add_argument('user_name', type=str, help='User Name', required=False)
     user_args.add_argument('chat_id', type=int, help='Chat ID', required=False)
-    user_args.add_argument('education_center', type=int, help='Education Center', required=False)
+    user_args.add_argument('education_center', type=str, help='Education Center', required=False)
     user_args.add_argument('verification', type=int, help='Verification', required=False)
     user_args.add_argument('payment', type=str, help='Payment Information', required=False)
 
@@ -43,7 +44,7 @@ user_resource_fields = {
     'name': fields.String,
     'user_name': fields.String,
     'chat_id': fields.Integer,
-    'education_center': fields.Integer,
+    'education_center': fields.String,
     'verification': fields.Integer,
     'payment': fields.String,
     'created_at': fields.DateTime(dt_format='iso8601'),
@@ -53,7 +54,10 @@ user_resource_fields = {
 
 class UserQueries(Resource):
     @marshal_with(user_resource_fields)
-    def get(self, chat_id):
+    @authentication_required
+    def get(self):
+        # Access user details from the request object
+        chat_id = request.chat_id
         # Get the User instance by ID
         user = User.query.filter_by(chat_id=chat_id).first()
 
@@ -90,14 +94,12 @@ class UserQueries(Resource):
         return user, 201
 
     @marshal_with(user_resource_fields)
-    def put(self, chat_id):
+    @authentication_required
+    def put(self):
         # Parse the arguments from the request
         args = update_user_args().parse_args()
-        print(f"Received PUT request with data: {args}")
-
-        # Get the existing User instance from the database
+        chat_id = request.chat_id
         user = User.query.filter_by(chat_id=chat_id).first()
-
         # Check if the instance exists
         if not user:
             abort(404, message="User not found")
@@ -112,8 +114,13 @@ class UserQueries(Resource):
         db.session.commit()
         return user
 
-    def delete(self, user_id):
-        user = User.query.filter_by(id=user_id).first()
+    @authentication_required
+    def delete(self):
+        chat_id = request.chat_id
+        if chat_id is None:
+            abort(400, message="Invalid user details")
+        user = User.query.filter_by(chat_id=chat_id).first()
+
         if user:
             db.session.delete(user)
             db.session.commit()
